+++ src/lists.h
@@ -22,9 +22,9 @@
 #include <mastodont.h>
 #include "session.h"
 
-char* construct_list(struct mstdnt_list* list, int* size);
+char* construct_list(struct mstdnt_list* list, size_t* size);
 char* construct_lists(struct mstdnt_list* lists, size_t size, size_t* ret_size);
-char* construct_lists_view(char* lists_string, int* size);
+char* construct_lists_view(char* lists_string, size_t* size);
 void content_lists(struct session* ssn, mastodont_t* api, char** data);
 
 #endif // LISTS_H
